requirements.txt
flask==2.3.3
python-hcl2==4.3.2
pyyaml==6.0.1
pygithub==1.59.1
python-gitlab==4.1.1
hvac==1.2.1
boto3==1.28.85
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
pytest==7.4.3
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
flask-cors==4.0.0

